+++ hypermedia/types/types.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from typing import TYPE_CHECKING, TypeAlias, TypeVar, Union
 
 from typing_extensions import Never, TypeVarTuple
@@ -15,7 +16,7 @@ class SafeString(str):
 NoChildren: TypeAlias = Never
 """Type alias for elements that are not allowed to have children."""
 
-PrimitiveChildren: TypeAlias = str | bool | int | float
+PrimitiveChildren: TypeAlias = str | bool | int | float | Decimal
 """Type alias for elements that are allowed to have only primitive children.
 
 Primitive children are ``str``, ``bool``, ``int`` and ``float``.